Added video progress bar and separated JavaScript
The JavaScript has been separated, made to have page specific functions to be in their own file. All shared functions between pages are now in shared.js.
Added files to be able to copy-paste sections of HTML to add sections like the control bar, playlist grid, and so on.
An mp3 file has been added to be able to test with.